Created tutorial on how to connect to UIC WiFi via Linux heavily based off of this article: https://help.uillinois.edu/TDClient/37/uic/KB/ArticleDet?ID=915
Lists all the fields to fill in both as a list and a table - will likely prune later.
Additionally: I need to add an up to date image of the Ubuntu Network Manager GUI as an example and to test this to ensure the tutorial is actually correct.
(It should be, but I'd rather ensure this is true before pushing it to main :) )