Incorporated Jacob Cohen's awk script!
Before this script, I was filtering out whitespace in a very...
interesting way. I was abusing the processing of arguments awk uses to
remove whitespace instead of properly and explicitly filtering for it.
Even if it was fine it was absolutely nasty on principle. Now it is
working without sacrificing our mortal souls. Beautiful.